pyproject.toml
[project]
name = "parker"
version = "0.1.0"
description = "Parker: A tool to dig some gold in upwork"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "aiohttp>=3.10.10",
    "python-dotenv>=1.0.1",
    "ruff>=0.7.1",
    "selectolax>=0.3.25",
]
